Initial implementation of a HTML/Javascript times table quiz
Pick the number of questions and how long per question and it will generate
unique questions, and start a countdown.

Press the Start button to begin.

Focus is given to each question in turn.
A correct answer is shown in green and a tick.
An incorrect or timed out answer is shown in red with a cross.